Repeat Answer Fix
Lines 65-68 I added an if and while stmt to reassign a random int from a_list.size to j until it didn't equal a_index. I tested with a smaller list of answers (4) to check, and I never got any repeats of the correct answer. Let me know if it doesn't work for you guys